FEAT [MULTI] : complete game creation

+++ lib/auth/data/models/ws_user_response.dart
@@ -293,7 +293,7 @@ class WsAnsweredQuestions extends Equatable {
         id: map?['qId'],
         level: map?['niveau'],
         isRightAnswer: map?['valid'],
-        time: (map?['time']).toDouble(),
+        time: (map?['time']),
       );
     } catch (e) {
       e.log(tag: 'WsAnsweredQuestions.fromMap');
